Add embed message author when using web app

+++ server/src/api/production.ts
@@ -1,6 +1,7 @@
 import { Express } from 'express';
 import { Server } from 'socket.io';
 
+import { atob } from '@global/utils';
 import { ProductionSchema } from '@schemas/productions.schema';
 import { IProdFinish } from '@global/interfaces/production.interface';
 import { CProductions } from '@interfaces/production.class';
@@ -19,18 +20,20 @@ export class ProductionHttp {
 
         // Add prod
         this._app.put(`${this._urlBase}/add`, (request, response) => {
+            const userId: string = request.query.userId as string;
             const prod: CProductions = new CProductions(request.body);
 
-            this._productionSchema.add(prod).then(() => {
+            this._productionSchema.add(prod, atob(userId)).then(() => {
                 response.send(prod);
             }).catch((err) => response.send(err));
         });
 
         // Delete prod
         this._app.delete(`${this._urlBase}/delete`, (request, response) => {
+            const userId: string = request.query.userId as string;
             const prod: CProductions = new CProductions(JSON.parse(request.query.value as string));
 
-            this._productionSchema.deleteById(prod).then(() => {
+            this._productionSchema.deleteById(prod, undefined, atob(userId)).then(() => {
                 response.send(prod);
             }).catch((err) => response.send(err));
         });
@@ -55,9 +58,10 @@ export class ProductionHttp {
 
         // Finish prod
         this._app.post(`${this._urlBase}/finish`, (request, response) => {
+            const userId: string | undefined = request.body.userId;
             const prod: CProductions = new CProductions(request.body);
 
-            this._productionSchema.finishProd(prod).then((finishProd: IProdFinish) => {
+            this._productionSchema.finishProd(prod, atob(userId)).then((finishProd: IProdFinish) => {
                 response.send(finishProd);
             }).catch((err) => response.send(err));
         });
